Changed node DOF Map and added BC setup
-- Node DOF map now contains all DOF Ids for each node, not just the
first DOF Id. Originally you would have had to get the node DOF set to
find out the size and iterate to get all DOFs for a particular node. Now
they are simply available in the NDM.

-- Dirichlet BC setup now in model.  The two vectors 'is_dof_fixed' and
'dof_value' tell you if a DOF is given a value, and then what that value
is. They are both 'ndof' in size.