Modified assembly to include DBC
-- Two copies of the DOF maps are kept. The normal version maps every
DOF to a unique ID. The ###_bc versions map only free DOFs. Fixed DOFs
are mapped to DOF::INVALID_ID. These DOFs are interpreted in the scatter
and gather routines as DOFs that need to be removed from the system.

-- The general gather and scatter routines have been replaced by
specialized versions specifically for different use cases (displacement,
force, stiffness matrices). This is because they need to do additional
things that would be impractical to do in general "scatter" and "gather"
routines.

-- a count of fixed DOFs and solvable DOFs is also stored in the model
now